requirements.txt
# Dependências para o script de atualização de certificados OpenVPN
PyYAML>=6.0          # Para leitura do arquivo config.yml
